+++ android/app/src/main/kotlin/co/za/netstart/mylist2/MainActivity.kt
@@ -1,0 +1,5 @@
+package co.za.netstart.mylist2
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()